refactor: JWT 관련 공통 코드 분리
